+++ main.go
@@ -9,6 +9,7 @@ import (
 	"os"
 
 	"github.com/opiproject/godpu/cmd"
+	"github.com/opiproject/godpu/cmd/ipsec"
 	"github.com/opiproject/godpu/cmd/network"
 	"github.com/opiproject/godpu/cmd/storage"
 	"github.com/spf13/cobra"
@@ -38,7 +39,7 @@ func newCommand() *cobra.Command {
 		},
 	}
 	c.AddCommand(cmd.NewInventoryCommand())
-	c.AddCommand(cmd.NewIPSecCommand())
+	c.AddCommand(ipsec.NewIPSecCommand())
 	c.AddCommand(storage.NewStorageCommand())
 	c.AddCommand(network.NewEvpnCommand())
 	return c
